Common working language for header and footer
