+++ blocks/footer/footer.scss
@@ -11,6 +11,7 @@
 }
 
 .footer {
+  --language-selector-display: block;
   margin: 0 auto;
   max-width: var(--non-spectrum-max-width);
   padding: 0;
