+++ scripts/language.js
@@ -1,3 +1,6 @@
+import { loadCSS } from './lib-franklin.js';
+import { htmlToElement } from './scripts.js';
+
 let languagePromise = null;
 export const loadLanguageFragment = async () => {
   languagePromise =
@@ -9,6 +12,7 @@ export const loadLanguageFragment = async () => {
           resolve(text);
         });
     });
+  loadCSS('/styles/language.css');
   return languagePromise;
 };
 
@@ -24,6 +28,13 @@ export const isIndex = () => {
   return url === '/';
 };
 
+export const getLanguagePath = (language) => {
+  if (isIndex()) {
+    return `/${language}`;
+  }
+  return window.location.pathname.replace(getCurrentLanguage(), language);
+};
+
 export const switchLanguage = async (language) => {
   const currentLanguage = getCurrentLanguage();
   if (currentLanguage === language) {
@@ -33,7 +44,47 @@ export const switchLanguage = async (language) => {
   if (isIndex()) {
     window.location.pathname = `/${language}`;
   } else {
-    const newPath = window.location.pathname.replace(currentLanguage, language);
-    window.location.pathname = newPath;
+    window.location.pathname = getLanguagePath(language);
   }
+};
+
+export const buildLanguagePopover = async (position) => {
+  const popoverId = 'language-picker-popover';
+  const popoverClass =
+    position === 'top' ? 'language-selector-popover language-selector-popover--top' : 'language-selector-popover';
+  let languagesEl = htmlToElement(await loadLanguageFragment());
+  languagesEl = languagesEl.querySelector('ul');
+
+  const languageOptions = languagesEl?.children || [];
+  const languages = [...languageOptions].map((option) => ({
+    title: option.textContent,
+    lang: option?.firstElementChild?.getAttribute('href'),
+  }));
+
+  const currentLang = getCurrentLanguage();
+  const options = languages
+    .map((option) => {
+      const lang = option.lang?.toLowerCase();
+      const selected = currentLang === lang ? 'selected' : '';
+      return `<span class="language-selector-label" data-value="${lang}" ${selected}>${option.title}</span>`;
+    })
+    .join('');
+  const popover = htmlToElement(`
+    <div class="${popoverClass}" id="${popoverId}">
+      ${options}
+    </div>`);
+
+  popover.addEventListener('click', (e) => {
+    const { target } = e;
+    if (target.classList.contains('language-selector-label')) {
+      target.setAttribute('selected', 'true');
+      const lang = target.getAttribute('data-value');
+      switchLanguage(lang);
+    }
+  });
+
+  return {
+    popover,
+    languages,
+  };
 };
